Restructure the UI related code.
